need to handle -0000001 type as valid number, 00001 is already fixed

+++ Libft/ft_atoi.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 
-long long	ft_atoi(const char *str)
+int	ft_atoi(const char *str)
 {
 	int	i;
 	int	sign;
